Began creating database handler
This script will handle the creation of the SQLite database used by the app. Currently, it runs when it is imported, but I will research how to have the script be run when Django starts. My intention is for the script to establish a connection to a database, and initialize a table from the csv file, and then create views for whatever purposes we require. Additionally, I have specified that the project uses SQLite as its SQL dialect.