Cleaned up README.md
I ended up splitting a number of lines to keep everything readable, as well as fixed up a few typos and slightly awkward language. Most of these changes won't be reflected in the browser view, as it's mainly to make it easier to read while editing. In addition, I have included a link to the "Our World in Data" website.