Reworked path construction for prototype graph setup
First of all, it's usually a good idea to open files using the "with ... as" clause, as it ensures that the file is closed when you're done. Additionally, the filenames for the images can be created by concatenating the strings together inside of parentheses, allowing you to skip having to assign a variable that you'll only use once anyways.