Removed sourcedata folder
The files included in this folder have been rendered redundant by the database, and are no longer needed. Any references to these files should probably be redirected to the input folder at the very least, as that's where the live csv file is kept, but they should ideally be making calls to the database, as the data there is automatically cleaned up and updated.