Created Dockerfile
This dockerfile will be used to run the server in a container. In order to get it to work, however, I also needed to add "localhost" to the list of allowed hosts. I have also created a .dockerignore file that currently ignores most of what we have ignored in the git repository. I may end up tweaking the specifics later down the road.