Added sqlite3 database to dockerignore file
I'm not entirely sure why it's not matching on the earlier one, but the database file should now definitely be excluded by the docker file. However, now this means that we'll likely have to make sure that we don't rely on the database persisting between runs, which will probably be less than ideal. I will have to look into the "testserver" option.