Overhauled graph generation
In order to avoid errors occurring due to certain countries missing months, I had to completely rethink how I implemented the queries used to collect the data. First, I created two new tables to pre-calculate the monthly data, as well as a collection of months that appear throughout the data. Next, I had to create a list comprehension that would make sure there were exactly as many months as we expect in the list. Finally, I had to create dummy data to fill in the missing months. All that's left for now is to clean up the user-facing aspects, such as the title getting too long to fit on the graph, and giving the bar graph a more descriptive title overall.