Changed database updater to only use valid columns
This is mainly to keep it from crashing whenever new columns are added to the site. Of course, we should try to have our models accurately reflect all of the data given, but for now, this will be a sufficient fail-safe mechanism to avoid failure when the data structure changes suddenly.