Added "Populate Data" command
This command will be used to update the database with data from OWID's github. In practice, we will probably want to stream the data directly into pandas, rather than saving it locally. However, there are also potential upsides to saving a local copy for quick testing purposes. A potentially more economical solution would be to check the timestamp file that is also present in the repo, as that would help check whether our version is out of date before potentially attempting an update that wouldn't affect anything.