Re-used received form, and added table
The table is currently rendered rather primitively, but it should be good enough as a simple implementation. I have also discovered that by simply passing the received form to the render function, the boxes will already be checked, without me having to do anything. As such, the custom widgets are not currently necessary, although I may still use them for other purposes, such as adding scripts to them.