Removed initialization container from compose file
Since we now have a command to handle the population of the database, we can simply run that command instead of relying on a bunch of containers being perfectly coordinated. As such, we no longer need to have a separate initialization container, as long as we can properly execute the migration/population from outside of the compose file.