Added environment variable to determine where the server is running
This should make it so that our code can operate differently based on whether or not the server is running on docker compose, google app engine, or on our local computer. Currently, the server doesn't seem to run well in a standalone container, but that probably won't matter that much.