Restructured template files
The template files have been separated by app, as well as placed into namespaces. In order to facilitate the use of the base template as a project-level template, I have moved it and the "about" page into a template directory located on the project root. Additionally, I set up the IDE to recognize the new template folder, as well as adding it to the list of directories scanned by the template engine in the settings file.