Began utilizing environment variables
The secret key will no longer be included in version control, and will either be stored as an environment variable, or as a secret. I will also migrate the reference to the secrets into app.yaml's environment variables, in order to decouple the project name from the code, as well as to potentially allow for different secrets to be accessed depending on whether the site is being run locally or on the cloud.