Updated docker configuration
In order to make it a bit easier to test using Docker Compose, I have modified the dockerfile to use a volume mount, allowing changes to the code on the host to be propagated to the container. I have also taken the liberty of updating the environment variables to help the server run properly in docker.