Added command to generate parquet
This command will fetch data from the source, and then convert it into a parquet format. This code is relatively inefficient in exchange for more control over the order of categories. This should only need to be run once a day, and it is identical to the code used for the daily update function.