Added data view class
Honestly, this probably didn't need to be a class, but the main difference this view makes is the fact that it uses pandas to directly plot the graphs. The graphs are currently converted into html through mpld3, but I'm looking into utilizing Bokeh. There's just a couple snags I need to iron out.