Reworked filtering, and moved title definition
Hopefully, the code should now be able to avoid looking at unnecessary files when handling requests. I had to rearrange some of the logic to allow the computations to work, but it should now work a bit better. Additionally, the title is no longer constructed using a list of country names, but rather, it uses data retrieved from the parquet files, which it was already looking at.